Vector Control, Events, Contact
Vector Control is now called in two parts: VectorControl1 and VectorControl2
VectorControl1 calls Bed_Net_1 and IRS_1
VectorControl2 calls Bed_Net_2 and IRS_2
The definitions are in bednet.R and irs.R

For fitting / scenario planning, all events are stored on the events object, and the `multiround` cases use it.

Functions to model Contact (or effective contact in relation to effective coverage) have been added for IRS and for Bed Nets.