shocks, irs_profiles, bednet_profiles
Options were added to modify `eir` models to configure F_shocks for multi-round.

Rather than expand functions for every profile, the shape parameters for multiround have been stored in data-raw and data/ and are accessible as irs_profiles and bednet_profiles.

Bednets no longer follow the IRS model -- the coverage profile and the net type are not linked.

event_length is now a prameter in the events

get_events functions added

pw is now passed with a default value of 1. Functions to store and pass pw are there.